Move interceptor file
